Add import-fails code as a fixture
